requirements.txt
numpy
tensorflow
tensorflow_hub
Pillow
fastapi
uvicorn
gunicorn==19.9.0
uvloop
httptools
scikit-learn